improve UART-read behaviour
The current behaviour was reading the input almost byte-by-byte,
leading to a flood of samples on the output port (and a lot of
inefficiencies). This modifies the behaviour to read each UART
exactly once per call to updateHook. The period at which the
task is deployed should therefore be lower than the time needed
to fill the 256 bytes of the sch-remote internal buffer. Note
that there is a detection of such overflows.